use one vector operator

+++ src/AtomicSupport.h
@@ -16,16 +16,16 @@ struct AtomicProposal
     unsigned nChanges;
 
     uint64_t pos1;
-    double delta1;
+    float delta1;
 
     uint64_t pos2;
-    double delta2;
+    float delta2;
 
-    AtomicProposal(char l, char t, uint64_t p1, double m1)
+    AtomicProposal(char l, char t, uint64_t p1, float m1)
         : label(l), type(t), nChanges(1), pos1(p1), delta1(m1), pos2(0), delta2(0.0)
     {}
 
-    AtomicProposal(char l, char t, uint64_t p1, double m1, uint64_t p2, double m2)
+    AtomicProposal(char l, char t, uint64_t p1, float m1, uint64_t p2, float m2)
         : label(l), type(t), nChanges(2), pos1(p1), delta1(m1), pos2(p2), delta2(m2)
     {}
 };
@@ -42,12 +42,12 @@ public:
     char mLabel;
 
     // storage of the atomic domain
-    std::map<uint64_t, double> mAtomicDomain;
+    std::map<uint64_t, float> mAtomicDomain;
     uint64_t mNumAtoms;
     uint64_t mMaxNumAtoms;
 
     // total mass of all atoms
-    double mTotalMass;
+    float mTotalMass;
 
     // matrix information
     uint64_t mNumRows;
@@ -56,10 +56,10 @@ public:
     uint64_t mBinSize;
 
     // average number of atoms per bin (must be > 0)
-    double mAlpha;
+    float mAlpha;
 
     // expected magnitude of each atom (must be > 0)
-    double mLambda;
+    float mLambda;
 
     // convert atomic position to row/col of the matrix
     uint64_t getRow(uint64_t pos) const;
@@ -76,13 +76,13 @@ public:
     AtomicProposal proposeExchange() const;
 
     // update the mass of an atom, return the total amount changed
-    double updateAtomMass(char type, uint64_t pos, double delta);
+    float updateAtomMass(char type, uint64_t pos, float delta);
 
 public:
 
     // constructor
-    AtomicSupport(char label, uint64_t nrow, uint64_t ncol, double alpha=1.0,
-        double lambda=1.0);
+    AtomicSupport(char label, uint64_t nrow, uint64_t ncol, float alpha=1.0,
+        float lambda=1.0);
 
     // create and accept a proposal
     AtomicProposal makeProposal() const;
@@ -92,15 +92,15 @@ public:
     MatrixChange getMatrixChange(const AtomicProposal &prop) const;
 
     // getters
-    double alpha() const {return mAlpha;}
-    double lambda() const {return mLambda;}
-    double totalMass() const {return mTotalMass;}
+    float alpha() const {return mAlpha;}
+    float lambda() const {return mLambda;}
+    float totalMass() const {return mTotalMass;}
     uint64_t numAtoms() const {return mNumAtoms;}
-    double at(uint64_t loc) const {return mAtomicDomain.at(loc);}
+    float at(uint64_t loc) const {return mAtomicDomain.at(loc);}
 
     // setters
-    void setAlpha(double alpha) {mAlpha = alpha;}
-    void setLambda(double lambda) {mLambda = lambda;}
+    void setAlpha(float alpha) {mAlpha = alpha;}
+    void setLambda(float lambda) {mLambda = lambda;}
     //void setMaxNumAtoms(uint64_t max) {mMaxNumAtoms = max;}
 };
